updating frontend to use new metricconfig
