+++ frontend/src/data/config/MetricConfigSDOH.ts
@@ -7,21 +7,15 @@ import {
   populationPctShortLabel,
   populationPctTitle,
 } from './MetricConfigUtils'
-import { SHOW_GUN_VIOLENCE } from '../providers/GunViolenceProvider'
-
-console.log('Gun violence preview enabled')
 
 export const SDOH_CATEGORY_DROPDOWNIDS = [
   'avoided_care',
   'health_insurance',
-  ...(SHOW_GUN_VIOLENCE ? ['gun_violence', 'gun_violence_youth'] : []),
   'poverty',
   'preventable_hospitalizations',
 ] as const
 
 export type SDOHDataTypeId =
-  | 'gun_violence_homicide'
-  | 'gun_violence_suicide'
   | 'poverty'
   | 'health_insurance'
   | 'preventable_hospitalizations'
@@ -31,21 +25,6 @@ export type SDOHMetricId =
   | 'ahr_population_pct'
   | 'avoided_care_pct_rate'
   | 'avoided_care_pct_share'
-  | 'fatal_population_pct'
-  | 'fatal_population'
-  | 'gun_violence_homicide_estimated_total'
-  | 'gun_violence_homicide_per_100k'
-  | 'gun_violence_homicide_pct_relative_inequity'
-  | 'gun_violence_homicide_pct_share'
-  | 'gun_violence_suicide_estimated_total'
-  | 'gun_violence_suicide_per_100k'
-  | 'gun_violence_suicide_pct_relative_inequity'
-  | 'gun_violence_suicide_pct_share'
-  | 'gun_violence_youth_deaths_estimated_total'
-  | 'gun_violence_youth_deaths_per_100k'
-  | 'gun_violence_youth_deaths_pct_relative_inequity'
-  | 'gun_violence_youth_deaths_pct_share'
-  | 'non_fatal_population_pct'
   | 'poverty_count'
   | 'poverty_estimated_total'
   | 'poverty_pct_rate'
@@ -55,16 +34,12 @@ export type SDOHMetricId =
   | 'poverty_population_pct'
   | 'preventable_hospitalizations_per_100k'
   | 'preventable_hospitalizations_pct_share'
-  | 'non_fatal_population_pct'
-  | 'non_fatal_population'
   | 'uninsured_estimated_total'
   | 'uninsured_pop_estimated_total'
   | 'uninsured_pct_rate'
   | 'uninsured_pct_relative_inequity'
   | 'uninsured_pct_share'
   | 'uninsured_population_pct'
-  | 'youth_population'
-  | 'youth_population_pct'
 
 export const UNINSURANCE_METRICS: DataTypeConfig[] = [
   {
@@ -288,204 +263,4 @@ export const PREVENTABLE_HOSP_METRICS: DataTypeConfig[] = [
       },
     },
   },
-]
-
-export const GUN_VIOLENCE_METRICS: DataTypeConfig[] = [
-  {
-    categoryId: 'sdoh',
-    dataTableTitle: 'Breakdown summary of gun homicides amongst adults',
-    dataTypeId: 'gun_violence_homicide',
-    dataTypeShortLabel: 'Homicides',
-    definition: {
-      text: 'Deaths caused by firearms used with the intent to harm others.',
-    },
-    description: {
-      text: 'Homicide fatalities by guns are a critical public health issue, as they often occur in contexts of other societal issues like poverty, inequality, and limited access to mental health services. Focusing on reducing gun-related homicides is essential for advancing health equity, especially in vulnerable populations.',
-    },
-    fullDisplayName: 'Gun homicides',
-    fullDisplayNameInline: 'gun homicides',
-    mapConfig: defaultHigherIsWorseMapConfig,
-    metrics: {
-      sub_population_count: {
-        chartTitle: '',
-        metricId: 'fatal_population',
-        shortLabel: 'Total Population for Gun Homicide Rates',
-        type: 'count',
-      },
-      pct_relative_inequity: {
-        chartTitle:
-          'Historical relative inequity of gun homicides amongst adults',
-        metricId: 'gun_violence_homicide_pct_relative_inequity',
-        shortLabel: '% relative inequity',
-        type: 'pct_relative_inequity',
-      },
-      pct_share: {
-        chartTitle: 'Share of total gun homicides amongst adults',
-        columnTitleHeader: 'Share of total gun homicides amongst adults',
-        metricId: 'gun_violence_homicide_pct_share',
-        populationComparisonMetric: {
-          chartTitle:
-            'Population vs. distribution of total gun homicides amongst adults',
-          columnTitleHeader: populationPctTitle,
-          metricId: 'fatal_population_pct',
-          shortLabel: populationPctShortLabel,
-          type: 'pct_share',
-        },
-        shortLabel: '% of homicides',
-        type: 'pct_share',
-      },
-      per100k: {
-        chartTitle: 'Rates of gun homicides amongst adults',
-        columnTitleHeader: 'Gun homicides per 100k people',
-        metricId: 'gun_violence_homicide_per_100k',
-        shortLabel: 'homicides per 100k',
-        trendsCardTitleName: 'Rates of gun homicides amongst adults over time',
-        type: 'per100k',
-        rateNumeratorMetric: {
-          chartTitle: '',
-          metricId: 'gun_violence_homicide_estimated_total',
-          shortLabel: 'Gun homicides',
-          type: 'count',
-        },
-        rateDenominatorMetric: {
-          chartTitle: '',
-          metricId: 'fatal_population',
-          shortLabel: 'Total Population',
-          type: 'count',
-        }
-      },
-
-    },
-  },
-  {
-    categoryId: 'sdoh',
-    dataTableTitle: 'Breakdown summary for gun suicides amongst adults',
-    dataTypeId: 'gun_violence_suicide',
-    dataTypeShortLabel: 'Suicides',
-    definition: {
-      text: 'Deaths resulting from individuals using firearms to inflict self-harm.',
-    },
-    description: {
-      text: 'Gun suicides represent a significant portion of firearm deaths. Addressing this issue is vital for health equity as it involves improving mental health services, access to care, and community support systems, particularly in areas with limited resources.',
-    },
-    fullDisplayName: 'Gun suicides',
-    fullDisplayNameInline: 'gun suicides',
-    mapConfig: defaultHigherIsWorseMapConfig,
-    metrics: {
-      sub_population_count: {
-        chartTitle: '',
-        metricId: 'fatal_population',
-        shortLabel: 'Total Population for Gun Suicide Rates',
-        type: 'count',
-      },
-      pct_relative_inequity: {
-        chartTitle:
-          'Historical relative inequity of gun suicides amongst adults',
-        metricId: 'gun_violence_suicide_pct_relative_inequity',
-        shortLabel: '% relative inequity',
-        type: 'pct_relative_inequity',
-      },
-      pct_share: {
-        chartTitle: 'Share of total gun suicides amongst adults',
-        columnTitleHeader: 'Share of total gun suicides',
-        metricId: 'gun_violence_suicide_pct_share',
-        populationComparisonMetric: {
-          chartTitle:
-            'Population vs. distribution of gun suicides amongst adults',
-          columnTitleHeader: populationPctTitle,
-          metricId: 'fatal_population_pct',
-          shortLabel: populationPctShortLabel,
-          type: 'pct_share',
-        },
-        shortLabel: '% of suicides',
-        type: 'pct_share',
-      },
-      per100k: {
-        chartTitle: 'Rates of gun suicides amongst adults',
-        columnTitleHeader: 'Gun suicides per 100k people',
-        metricId: 'gun_violence_suicide_per_100k',
-        shortLabel: 'suicides per 100k',
-        trendsCardTitleName: 'Rates of gun suicides amongst adults over time',
-        type: 'per100k',
-        rateNumeratorMetric: {
-          chartTitle: '',
-          metricId: 'gun_violence_suicide_estimated_total',
-          shortLabel: 'Gun suicides',
-          type: 'count',
-        },
-        rateDenominatorMetric: {
-          chartTitle: '',
-          metricId: 'fatal_population',
-          shortLabel: 'Total Population',
-          type: 'count',
-        }
-      },
-    },
-  },
-]
-
-export const GUN_VIOLENCE_YOUTH_METRICS: DataTypeConfig[] = [
-  {
-    categoryId: 'sdoh',
-    dataTableTitle: 'Breakdown summary of gun deaths among youth',
-    dataTypeId: 'gun_violence_youth',
-    dataTypeShortLabel: 'Gun Deaths',
-    definition: {
-      text: 'Deaths of individuals under the age of 26 caused by firearms.'
-    },
-    description: {
-      text: 'Measuring gun deaths among youth is crucial because it helps us understand the impact of firearm violence on younger populations, guiding the development of targeted interventions and policies to protect our most vulnerable citizens and prevent future tragedies.'
-    },
-    fullDisplayName: 'Gun deaths',
-    fullDisplayNameInline: 'gun deaths',
-    mapConfig: defaultHigherIsWorseMapConfig,
-    metrics: {
-      sub_population_count: {
-        chartTitle: '',
-        metricId: 'gun_violence_youth_deaths_estimated_total',
-        shortLabel: 'Total Population for Gun Deaths (Youth)',
-        type: 'count',
-      },
-      pct_relative_inequity: {
-        chartTitle: 'Historical realative inequity of gun deaths among youth',
-        metricId: 'gun_violence_youth_deaths_pct_relative_inequity',
-        shortLabel: '% relative inequity',
-        type: 'pct_relative_inequity',
-      },
-      pct_share: {
-        chartTitle: 'Share of total gun deaths among youth',
-        columnTitleHeader: 'Share of total gun deaths among youth',
-        metricId: 'gun_violence_youth_deaths_pct_share',
-        populationComparisonMetric: {
-          chartTitle: 'Population vs. distribution of total gun deaths among youth',
-          columnTitleHeader: `${populationPctTitle} (ages 0-25)`,
-          metricId: 'population_pct',
-          shortLabel: populationPctShortLabel,
-          type: 'pct_share'
-        },
-        shortLabel: '% of gun deaths',
-        type: 'pct_share',
-      },
-      per100k: {
-        chartTitle: 'Rates of gun deaths among youth',
-        columnTitleHeader: 'Gun deaths among youth per 100k people',
-        metricId: 'gun_violence_youth_deaths_per_100k',
-        shortLabel: 'deaths per 100k',
-        trendsCardTitleName: 'Rates of gun deaths among youth over time',
-        type: 'per100k',
-        rateNumeratorMetric: {
-          chartTitle: '',
-          metricId: 'gun_violence_youth_deaths_estimated_total',
-          shortLabel: 'Gun deaths (youth)',
-          type: 'count',
-        },
-        rateDenominatorMetric: {
-          chartTitle: '',
-          metricId: 'youth_population',
-          shortLabel: 'Total Population',
-          type: 'count',
-        }
-      },
-    }
-  }
 ]
