+++ frontend/src/data/providers/GunViolenceProvider.ts
@@ -29,8 +29,6 @@ export const GUN_VIOLENCE_SUICIDE_METRICS: MetricId[] = [
 export const POPULATION_METRICS: MetricId[] = [
   'fatal_population_pct',
   'fatal_population',
-  'non_fatal_population_pct',
-  'non_fatal_population',
 ]
 
 const GUN_VIOLENCE_METRICS: MetricId[] = [
